requirements.txt
absl-py==2.1.0
astroid==3.3.5
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
dill==0.3.9
executing==2.1.0
flatbuffers==24.3.25
fonttools==4.55.0
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.12.1
idna==3.10
ipython==8.29.0
isort==5.13.2
jedi==0.19.2
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
parso==0.8.4
path==17.0.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==4.25.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydot==3.0.2
Pygments==2.18.0
pylint==3.3.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scipy==1.14.1
six==1.16.0
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tomlkit==0.13.2
traitlets==5.14.3
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
typing_extensions>=4.6.0 # modified
